\name{Conjecture}
\alias{Fig7}
\title{Conjecture}
\usage{
  Fig7(rho=0.5, fill=c("gray85", "gray70"))
}
\description{
  Illustrates the idea behind the proof of the conjecture given in the
  appendix.
}
\arguments{
  \item{rho}{Correlation between the two features.  Can be negative.}
  \item{fill}{Colors used to fill shaded regions.}
}
\examples{
  Fig7(0.5)
  Fig7(-0.5)
}
