\name{TCGA data analysis}
\alias{Fig6}
\title{TCGA data analysis}
\usage{
  Fig6(penalty="lasso")
}
\description{
  Analyzes the breast cancer TCGA data set; see \code{\link{loadData}}.
}
\arguments{
  \item{penalty}{Penalty to use; can be \code{lasso}, \code{MCP}, or
  \code{SCAD}.}
}
\examples{
  Fig6()
}
